requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
fastapi
uvicorn
openai
langchain_openai
langchain
dynaconf
pytest
python-dotenv
pydantic-settings
shortuuid

polyglot
pyicu
pycld2
pillow

langchain-community